move templates to their own file
